+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * New cop `EnsureReturn` tracks usages of `return` in `ensure` blocks.
 * New cop `HandleExceptions` tracks suppressed exceptions
 * New cop `AsciiIdentifiersAndComments` tracks uses of non-ascii characters in identifiers and comments.
+* New cop `AvoidLiteralStringArray` tracks uses of literay array syntax for array of strings.
 
 ### Bugs fixed